Use tableview.refresh and move custom cell
