+++ src/main/java/ledger/user_interface/ui_controllers/component/tablecolumn/event_handler/DateEventHandler.java
@@ -37,7 +37,7 @@ public class DateEventHandler implements EventHandler<TableColumn.CellEditEvent<
             Optional<ButtonType> result = alert.showAndWait();
 
             if(result.get() != ButtonType.OK){
-                ((TransactionTableView)t.getTableView()).updateTransactionTableView();
+                t.getTableView().refresh();
                 return;
             }
         }
