+++ src/main/java/ledger/user_interface/ui_controllers/component/tablecolumn/event_handler/PayeeEventHandler.java
@@ -37,7 +37,7 @@ public class PayeeEventHandler implements EventHandler<TableColumn.CellEditEvent
             if(accountToSet.equals(transaction.getAccount())){
                 setupErrorPopup("Source and Destination accounts can not be the same");
 
-                //((TransactionTableView)t.getTableView()).updateTransactionTableView();
+                t.getTableView().refresh();
                 return;
             }
             transaction.setTransferAccount(accountToSet);
